+++ src/triple/helpers.js
@@ -1,8 +1,12 @@
+import { createSelector } from 'reselect'
+
 import isArray from 'lodash/isArray'
 import isPlainObject from 'lodash/isPlainObject'
 import isString from 'lodash/isString'
+import mapValues from 'lodash/mapValues'
 
 import { isEntity, isEntityCreated } from '../entity/helpers'
+import { buildTypeIndex, entitySelector } from '../entity/select'
 
 export function isTriple(triple, checkCreated = false) {
   const { id, subject, predicate, object } = triple
@@ -18,4 +22,13 @@ export function isTriple(triple, checkCreated = false) {
     if (!isEntityCreated(object)) throw new Error('Triple object must have an id already.')
   }
   return true
+}
+export function expandIndex(items, entity) {
+  return mapValues(items, (val, id) => entity[id])
+}
+export function tripleIndexSelector(indexSelector) {
+  return createSelector(indexSelector, entitySelector, expandIndex)
+}
+export function tripleTypeIndexSelector(indexSelector) {
+  return createSelector(tripleIndexSelector(indexSelector), buildTypeIndex)
 }